Separate document import logic into modules
