+++ src/main/java/com/vnaskos/pdfguru/execution/document/DocumentControlListener.java
@@ -1,4 +1,4 @@
-package com.vnaskos.pdfguru.execution;
+package com.vnaskos.pdfguru.execution.document;
 
 import java.io.IOException;
 
@@ -7,4 +7,6 @@ public interface DocumentControlListener<DOCUMENT, PAGE> {
     DOCUMENT getOutputDocument();
 
     void addPage(PAGE page) throws IOException;
+
+    void preventFromGC(DOCUMENT document);
 }
