+++ src/test/java/com/vnaskos/pdfguru/execution/document/pdfbox/PdfboxDocumentManagerTest.java
@@ -1,5 +1,6 @@
-package com.vnaskos.pdfguru.execution;
+package com.vnaskos.pdfguru.execution.document.pdfbox;
 
+import com.vnaskos.pdfguru.execution.ExecutionProgressListener;
 import com.vnaskos.pdfguru.input.items.InputItem;
 import org.apache.pdfbox.pdmodel.PDDocument;
 import org.apache.pdfbox.pdmodel.PDPage;
@@ -13,7 +14,6 @@ import java.io.IOException;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.ArgumentMatchers.anyString;
 import static org.mockito.Mockito.*;
-import static org.mockito.Mockito.verify;
 
 public class PdfboxDocumentManagerTest {
